Lezione 17 - Visual - AI/gemini/requirements.txt
google-cloud-aiplatform
google-cloud-texttospeech